P2P Test for up to 12.8 GB/sec
This is based on the P2P Basic example for FlexRIO.  The Basic example can transfer up to 800 MB/sec.  Current NI PXIe FPGA products can transfer up to 7 GB/sec.  The Basic example has been modified to support up to 12.8 GB/sec.  This is by changing the transfer size from one (1) U64 to sixteen (16) U64s.  The clock rate in all is 100 MHz.  As with the Basic example, the example will 'run' at the maximum rate supported.  The committed code is for the hardware I have access to (7915 and 5785) which on a 1095 chassis and 8880 controller can transfer the specified 7 GB/sec with no errors.  This can be ported to other products which can support P2P rates of more than the 800 MB/sec supported by the existing example.